🚚 Rename: 폴더 컨벤션에 맞춰 폴더명 수정
